[anjali] refactored : improved the ui for creating a scheme by manually passing planId
