+++ src/app/components/admin/admin.component.html
@@ -1,1 +1,10 @@
-<app-plan-form></app-plan-form>
+<body>
+    <div class="dark-theme">
+      <div class="add-plan-box" routerLink="/dashboard/admin/add-plan">
+        <span>Add New Plan</span>
+      </div>
+      <div class="add-scheme-box" routerLink="/dashboard/admin/add-scheme">
+        <span>Add Scheme</span>
+      </div>
+    </div>
+</body>
