+++ src/app/components/admin/admin.component.scss
@@ -1,0 +1,32 @@
+body {
+    display: flex;
+    justify-content: flex-start;
+    align-items: center;
+    background-color: black;
+    color: white;
+  
+    .dark-theme {
+        margin-left: 30px;
+        display: flex;
+        gap: 10px;
+        
+        .add-plan-box,
+        .add-scheme-box {
+            background-color: #1e1e1e;
+            padding: 20px;
+            border-radius: 8px;
+            box-shadow: 0 4px 6px rgba(0, 0, 0, 0.6);
+            transition: background-color 0.3s ease;
+            cursor: pointer;
+    
+            span {
+                font-size: 18px;
+                text-align: center;
+            }
+    
+            &:hover {
+                background-color: #292929;
+            }
+        }
+    }
+}
