+++ src/app/components/admin/plan-form/plan-form.component.html
@@ -1,12 +1,30 @@
 <body>
-    <div class="dark-theme">
-      <div class="add-plan-box" (click)="openPlanDialog()">
-        <span>Add New Plan</span>
-      </div>
-      <div class="add-scheme-box" (click)="openSchemeDialog()">
-        <span>Add Scheme</span>
-      </div>
-    </div>
+  <div class="plan-form-container">
+      <h2>Create New Plan</h2>
+      <form [formGroup]="planForm">
+          <div>
+              <div class="form-group">
+                  <label for="planName">Plan Name*</label>
+                  <input id="planName" type="text" formControlName="planName" class="form-control">
+                  <span class="error" *ngIf="isFormSubmitted && planFormControls['planName']?.errors?.['required']">
+                      Plan Name is required
+                  </span>
+              </div>
+              <div class="form-group">
+                  <label for="description">Description (Optional)</label>
+                  <textarea id="description" formControlName="description" class="form-control"></textarea>
+              </div>
+              <div class="form-group">
+                  <label for="category">Category*</label>
+                  <input id="category" type="text" formControlName="category" class="form-control">
+                  <span class="error" *ngIf="isFormSubmitted && planFormControls['category']?.errors?.['required']">
+                      Category is required
+                  </span>
+              </div>
+          </div>
+          <br>
+          <button (click)="onClose()">Cancel</button>
+          <button class="createplanbutton" (click)="handleSubmit()">Create Plan</button>
+      </form>
+  </div>    
 </body>
-
-
