+++ src/app/components/admin/plan-form/plan-form.component.ts
@@ -1,7 +1,7 @@
 import { Component } from '@angular/core';
-import { MatDialog } from '@angular/material/dialog';
-import { PlanDialogComponent } from '../plan-dialog/plan-dialog.component';
-import { SchemeFormComponent } from '../scheme-form/scheme-form.component';
+import { FormBuilder, FormGroup, Validators } from '@angular/forms';
+import { Router } from '@angular/router';
+import { HttpService } from 'src/app/services/http-service/http.service';
 
 @Component({
   selector: 'app-plan-form',
@@ -9,21 +9,47 @@ import { SchemeFormComponent } from '../scheme-form/scheme-form.component';
   styleUrls: ['./plan-form.component.scss']
 })
 export class PlanFormComponent {
-  constructor(private dialog: MatDialog) {}
+  planForm!: FormGroup;
+  isFormSubmitted = false;
+  
+  constructor(private formBuilder: FormBuilder, private httpService: HttpService, private router: Router) { }
 
-  openPlanDialog(): void {
-    this.dialog.open(PlanDialogComponent, {
-      width: '400px',
-      height: 'auto',
-      disableClose: true
+  ngOnInit(): void {
+    this.planForm = this.formBuilder.group({
+      planName: ['', Validators.required],
+      description: [''],
+      category: ['', Validators.required]
     });
   }
 
-  openSchemeDialog(): void {
-    this.dialog.open(SchemeFormComponent, {
-      width: '800px',
-      height: 'auto',
-      disableClose: true
-    });
+  get planFormControls() {
+    return this.planForm.controls;
+  }
+
+  handleSubmit() {
+    this.isFormSubmitted = true;
+    if (this.planForm.valid) {
+      const { planName, description, category } = this.planForm.value;
+    
+      const payload: { planName: string; category: string; description?: string } = {planName, category};
+    
+      // Add description if it's not empty
+      if (description) {
+        payload.description = description;
+      }
+      this.httpService.postApiCall('/api/v1/plan', payload).subscribe({
+        next: (res) => {
+          console.log('Plan created successfully:', res);
+          this.router.navigate(['/dashboard/plans']); 
+        },
+        error: (error) => {
+          console.error('Error creating plan:', error);
+        },
+      });
+    }
+  }
+
+  onClose(): void {
+    this.router.navigate(['/dashboard/plans']); 
   }
 }
