+++ src/app/components/admin/scheme-form/scheme-form.component.scss
@@ -1,9 +1,10 @@
 body {
-    background-color: #121212;
+    background-color: black;
     color: white;
     display: flex;
     justify-content: center;
     align-items: center;
+    margin-top: 20px;
 }
   
 .scheme-form-container {
